add typing to controller, move interface into separate directory
