+++ server/controllers/appt.controller.ts
@@ -1,7 +1,8 @@
 import { Request, Response } from "express";
 import Appointment from "../models/appt.model";
+import { IAppointment } from "../interfaces/appointment.interface";
 
-const findAllAppointments = async (req: Request, res: Response) => {
+const findAllAppointments = async (req: Request, res: Response): Promise<IAppointment[]> => {
   try {
     const appointments = await Appointment.find()
     return res.status(200).send(appointments)
@@ -10,7 +11,7 @@ const findAllAppointments = async (req: Request, res: Response) => {
   }
 }
 
-const createNewAppointment = async (req: Request, res: Response) => {
+const createNewAppointment = async (req: Request, res: Response): Promise<IAppointment> => {
   try {
     const appointment = await Appointment.create(req.body)
     return res.status(201).send(appointment)
@@ -19,7 +20,7 @@ const createNewAppointment = async (req: Request, res: Response) => {
   }
 }
 
-const findOneAppointment = async (req: Request, res: Response) => {
+const findOneAppointment = async (req: Request, res: Response): Promise<IAppointment> => {
   try {
     const appointment = await Appointment.findOne({ _id: req.params.id })
     return res.status(200).send(appointment)
@@ -27,7 +28,7 @@ const findOneAppointment = async (req: Request, res: Response) => {
     return res.status(500).send(e.message)
   }
 }
-const updateAppointment = async (req: Request, res: Response) => {
+const updateAppointment = async (req: Request, res: Response): Promise<IAppointment> => {
   try {
     const appointment = await Appointment.findByIdAndUpdate({ _id: req.params.id }, req.body, { new: true })
     return res.status(200).send(appointment)
@@ -35,7 +36,7 @@ const updateAppointment = async (req: Request, res: Response) => {
     return res.status(500).send(e.message)
   }
 }
-const deleteAppointment = async (req: Request, res: Response) => {
+const deleteAppointment = async (req: Request, res: Response): Promise<{}> => {
   try {
     const appointment = await Appointment.deleteOne({ _id: req.params.id })
     return res.status(200).send(appointment)
@@ -44,4 +45,10 @@ const deleteAppointment = async (req: Request, res: Response) => {
   }
 }
 
-export { findAllAppointments, createNewAppointment, findOneAppointment, updateAppointment, deleteAppointment }
+export { 
+  findAllAppointments,
+  createNewAppointment,
+  findOneAppointment,
+  updateAppointment,
+  deleteAppointment
+}
